Updated Player.cs
I got some nice controls setup for both controller and keyboard.

So now, the player will walk unless the joystick is tilted to a certain distance.

The south button on the controller is always jump, so regardless of controller it's always in the same place.

I'll probably use UI buttons that don't have symbols on them and only show what the button does so it's just less confusing.

TODO:

Add in a "falling state" when you run out of jumps so you just start spinning like in smash.

Decide how you are going to do the attack input.

Add some scale animations so that there is an example animation state machine that reacts to movements and jumping and just scales the player accordingly.

Try to cut down on the amount of code and variables you've made. It seems like there's a lot and there's probably a way to decrease them and retain the same functionality.

Go over the comments and fix them so that they aren't guidelines for coding but instead just describe what the code  does.