Update Player.cs
I started setting up the more complex controls with attacks and the flipping of the player based on direction.

TODO:
Lots of things, I need to find a better way to make the player flip consistently and I think I need to change the dead zone for the controller bc it's only happening when I use an xbox controller and let the joystick fling back to the center.

I need to start setting up the game manager after I finish setting up the controls for the player so that I can actually create characters and then start testing gameplay.