Update Player.cs
I finally got a good method working for determining which way the player sprite should face relative to the player's input.

I just check if the player's input has been going in the same direction as the previous frame, if so then we change the direction it is currently facing to match input. That way if the player lets their control stick fling back it won't register that and switch directions, they would need to voluntarily keep pushing the stick farther.

TODO:

Make it so that the frame you press the  jump button there is a single frame after where the player does a charging animation before triggering the jump so that if they attempt to do an upsmash they have time to put in the input. You also may want to not do a frame-specific count but actually base it off of delta time. For now, I'm trying to figure out how to make jumping faster while maintaining the correct height.