Updated Player.cs
I'm going to change the control scheme setup to reduce the amount of code that exists and just set up a specific input combo for each ability. I don't know why I didn't do it before but now I feel kind of dumb for wasting my time with dot product as doing conditionals for attacks is much more efficient than doing dot product every frame we decide to attack. It'll reduce the size of the code and simplify it. This should be perfect.

TODO:
Change all the input for everything other than Smash attacks to resemble how smash attacks are done.

That way I get rid of a ton of code and just subscribe the method of the attack to its specific input.

I wasted lots of time, but realized something quite valuable so I'm counting this as a win.