Update Player.cs [IMPORTANT]
I started setting up individual controls in the InputManager for each attack type so I could just sub them to the method they were part of. Only problem is some attacks use the same combo of buttons with the only difference being that one has the control stick pushed all the way and the other isn't pushed all the way.

I need to make a control buffer if I want it to work properly. I'm stopping for the night but hopefully I can get this buffer working tomorrow so I can see what the player is actually trying to input.