Update Player.cs
I have made some significant changes and I'm going to go back to doing more complex if statements for controls.

The new input system isn't built for what I need it to do so I'll just do it in code.

I got a working check for if the player is tapping the left control stick in a direction. Only problem is that we still register a tilt attack if the player doesn't hit the attack button on the EXACT frame they tap the control stick.

In SB64 it's usually easier to perform a forward smash than it is to perform a forward tilt, because for the forward tilt to occur you have to be standing still, holding the control stick all the way in a direction, and then pressing the attack button. I don't know if that's correct tho as I'm using an emulator. I should've brought my switch back over break to test this stuff properly.

I'm uploading an incomplete changeset because I need to.

What I plan on is properly checking if the player is using "intent" with their joystick motions, and if they are we do a smash atttack, otherwise if they are sitting in the tilt range for tilt we should do a tilt attack.

Currently smash attacks are really inconsistent and tilt attacks are really easy to land.