Setup animations.
I got the jumping animation working, I'm going to refer to my code from knock it off socks to figure out when to play the landing animation. Though, I will say that the animations in knock it off socks are individual frames being swapped to match the speed of movement.