Update Player.cs
Ok, So I re-did the code for tapping because I realized it was wrong. Now, tapping the control stick is perfectly setup. Currently though, the window of time for tapping does not apply to smash attacks so you can just hit the attack button long after tapping and still do a smash attack. I'll fix it soon but I want to upload this before I start modifying it.

I also imported a sample that shows what the input looks like (a controller visualizer). Which is from the new input system package. There are some more samples in there I need to check out, especially if I want to be able to implement rebinding controls in the future or visualize said controls.

For now, I'll fix working on the smash attacks and the tilt attacks. The tilt attacks aren't registering now either.