Update Player.cs
Smash attacks and tilt attacks now seem to function properly. I think there are a couple of small inconsistencies with input and I'll work it out. I think I'm going to move on to doing some animations and such for the first character in the game and we'll see how it works. Hopefully, I can make a basic sprite and animation set for them and it'll work fine. I think I'll make norm the niner or something like that. I'll also have to make a sandbag or something to practice against with attacks.