Updated animations.
I fixed the facing direction issue and now I need to figure out why the WASD input doesn't allow for direction switching. Probably because it's values are normalized and not based on push distance like a stick is.

I also need to code the actual game stuff soon. I think I'll make it so that the player can have a custom death explosion effect/animation. I need to also do the animations for attacks for norm. We'll see how that goes.