Updated Sprites.
I changed the norm prefab so it uses the .aesprite file for its sprites and this also allows me to properly have other items be put in norm's hand in case I add an items system later. This also helps with his attacks that require a pickaxe for the animation. A normal metal pickaxe for tilt attacks and then a gold pickaxe for the smash attacks.