Started implementing attack animations.
I've started to implement the attacking animations for tilt attacks but I noticed that tapping isn't working properly again. I need to figure out why but for some reason it seems that it's back to the original issue where regardless of joystick speed the input gets registered as a tap so long as the player's input reaches 1 on any of the axes. I need to figure this out and then fix it. After I implement some animations for attacking I'm going to begin setting up multiple players so I can test attacking and input/controller management.