Updated animations
I got the animations for norm updated so now his up tilt works. The layers are finnicky so his running animation only plays during an attack if the attack layer weights are at 0,5 or lower. It's not a big problem and I'd rather continue working and look into it later.