Setting up Character Icons
I set up a lot of things with the UI in this one, I also wasted like 30 minutes animating the percentage numbers changing only to realize that incrementing them isn't a good animation effect and would just make the game feel slower by proxy. I'm going to upload quick with the animation then I'll remove it and upload again.

I also setup dynamically adding character icons to the UI when they join the game. I need to figure out how to do joining prior to loading the PVP scene so that  the players exist the second the level loads.

I'll also have to do a countdown at the beginning of the PVP level.

I set up a lot with player icons including making a template for them so that developers can copy it for use with their character.