Update PvPScene
I fixed an issue with the camera where it wasn't properly constrained to the edge of the map's area.

I made the killBounds using a polygon collider and I believe that putting a hole in the center of it worked. It's hard to tell if that rectangular shape in the center is a hole or if it counts as a filled part of the collider.

I'm going to do an upload after this because I will be removing some things that didn't work but I want to have the back up just in case.