Update PvPScene.unity
I fixed some issues with Player.cs because I accidentally setup it's raycasting mask wrong.

I finalized the size of the kill bounds and where the camera stops when the player moves off the screen. There is a small amount of space between the "edge of the map" and the "kill boundary" just like smash.

Now, I'm going to finish implementing the hurtboxes so that attacks will be possible.