Update Player.cs
I implemented the Smash Bros. Formula for knockback and it seems to work really well. Only thing is that I haven't fully implemented it yet and that in smash the launching visuals when a player goes flying look much slower than ours. I'll look into it but I'm guessing it has to do with the weight of the player being set to 1 in our game and apparently smash's default weight unit starts at 100 for their formula if it's set to ignore weight. I'll test it some more.