Update Player.cs
I've implemented the smash formula for knockback the best I can for now. It doesn't translate well because we aren't using a fully exposed physics engine like smash was.

it works for now but I'm getting kind of tired of the amount of time it takes to test this stuff. I really wish that I had more time.