Update Player.cs
In smash the units are 0.1 of a meter, therefore for the formula to be scaled properly we need to multiply by 10 to get the value scaled to how our game is setup.

I think I got the formula for knockback and launching correct it just isn't fully implemented yet.

Also, use MovePosition if the framerate is too inconsistent instead of rb.velocity = launchSpeed; in the LaunchCoroutine.

I also tried my best to cap the game at 60fps and cap deltaTime at 1/60th of a second but I probably shouldn't do that to deltaTime but for some reason capping the fps doesn't actually cap it so it can still get up to 70fps.

I need a formula that takes deltaTime and returns a 1 when deltaTime is at 1/60th of a second. Then, when it's above it gives a smaller value to scale properly with execution speed.

I have lots to fix still but know that I need to remove the deltaTime cap but keep the frameRate cap so that it works the way smash does (smash runs at 60fps).