Started Implementing the character select screen.
I got it working so when you press a button on the controller it spawns a cursor and the cursor prints out the object under it in the UI when you press the south button on any controller or click with the mouse. This works for now but there's still a lot to do and cleanup.

TODO:

I need to store whichever character they select in this scene in the game manager with the device that selected it then instantiate the prefab and assign the device to that specific prefab.