Continued implementing character select
Added scriptable objects to define character icons and allow them to be procedurally loaded.

Added a global list of all the character icons so that people can just add their icon to that list and it'll automatically be put on the selection screen.

TODO:

I need to make it so that when you select a character it drops the coin on the character and shows that character in the bottom area with their P1 or whatever player selected them, then when a button is selected we load into the PVP scene and assign the controller that selected the specific character to the newly instantated playerInput component. This is going to be a funky thing to design, I'll probably just store a dictionary of the devices and their prefab then just loop through and link the device to the newly instantiated prefab.

This will be difficult to figure out.