Completed the Character Selection implementation.
I completed the implemetation for character selection. There are probably some glitches in it right now but I'm satisfied with it.

TODO:
Make a game mode script and instead of adding the selected characters to the game manager add them to the game mode and store different data based on game mode type, for example 3 stock or 5 stock.