Update Player.cs
I am getting really close to having accurate launching but the gravity is the hardest part right now. I'm not sure how it's supposed to be applied.

I also added something for calculating hitstun for a launch because the amount of calculated frames during a launch is actually based off of the hitstun frame count.

TODO:
Implement hitstun state for the player.