Update Player.cs
I've managed to figure out the formula such that I can set the time and height of a jump and it is sped up and reaches the desired height. Only issue is that it isn't accurate. the height reached is like off by 1 which is an issue. I need to figure it out but I don't have time to fix it.