Smash-like jumping
I thought I pushed to this repo, but I didn't so here's what this push was supposed to say 6 hours ago.

I was able to achieve Smash-Like jumping on accident during GGJ 2024. I realized that, for whatever reason, Unity's interpolation was the issue. I am going to convert the code to only do physics in fixedUpdate.

I changed the physics timestep to 0.0166666666 which gets rounded to a 7 at the last digit. This makes fixedUpdate exactly 60fps.

OK, BEFORE I COMMITTED THIS I WENT BACK TO WORK ON JUMPING FOR A SECOND AND NOW IT'S LITERALLY PERFECT.

I took the desired height divided by the height we got and always got a constant of 1.2 at 60fps and 60 fixedUpdate iterations. I then multiply the height by 1.2 before passing it to the jump formula and it now perfectly, no matter what gets to the EXACT desired jump height. I am not kidding when I say that I am genuinely surprised that I have gotten this working.

Oh, also the jumping actually works like smash jumping now where you go up fast (timeToApex sets this) and fall at a normal rate (baseGravity).

I am so excited by this stuff because it means I can probably use this formula in 3D as well. This is crazy, like, I am genuinely so happy.

Next step: figure out how to calculate the constant based off timeToApex so that it works for all the different amounts of time.

Now I have to figure out how to undo the commit on the other repo.