Update Player.cs
I changed the code so that it calculates the direction we are hit in based off of the hurtbox's position and the hit player's position. This allows me to just automatically do direction code instead of having to check if the player is facing left then set the angle to be negative then checking for it later and changing another value.

What I need to do:
Change all the times I'm modifying the force or velocity to be in fixedUpdate, I didn't know but even if you have interpolate on you still aren't supposed to put addforce calls in update. I figured that was the point of it but I didn't realize it.

Other things I should do:

I should really really go through and optimize the player's code soon. It's pretty rough.

I also really need to make a nice readme then post about this on the game devs discord.