Update README.md
I had a realization about coroutines/classes that are a specific move. I may want to rethink the idea of how moves are currently designed and instead use a class containing a coroutine or a coroutine as a parameter for each move where the dev can specify their functionality then pass it. This might be overcomplex but it's just an idea. I think it would work better for things like dashing.