Update Player.cs
Dashing is much more consistent now as I've updated the coroutine. If you input a dash distance it makes it pretty close to that distance and will always travel the same distance every time you dash.

There is a way to make it dash the exact distance which is what the modifier is for.

If you set the position of the player to zero then dash to the right divide the desired distance you input by the actual x value reached and set the value you got to be the "Dash Modifier". It will now always reach your desired distance for that specific framerate.

I also need to fix some of the ground check code but I fixed it enough for now.

I need to fix some of the colliders for the sprites in the PvPscene

TODO: Add dash attacks.