Update Player.cs
I was messing around with the dashing and now it's really close to being accurate. It's still off by one value of it's calculated acceleration which is odd but I think it has something to do with how I have to apply a force then wait until the next fixedupdate for it to be calculated. I may need to call the yield function after applying force in the loop then decrement the number  of frames. We'll find out eventually but it's good enough so I'm going to work on some HW.

I still need to add a check for dash attacking and I need to implement dodging and shielding.