Update Player.cs [FIXED LAUNCHING]
I fixed the issue with launching, it had to do with how the isGrounded bool was coded, I switched it back to the old way and now it's perfect. I'll probably go in and delete that extra branch when I get a chance as it's no longer needed.

I kind of implemented Dash Attacks, so if your values for the moveset reset, that's my bad.

Dashing needs some improvement and for some reason holding left makes you dash instead of dashing, then running. I'll have to figure that out after this commit.