Update Player.cs
I refactored a lot of the variables in Player.cs but it still needs lots of work to not be considered spaghetti code.

I changed a lot of the order of parameters as I was updating the wiki and the setup instructions.

I also changed how the CharacterBase prefab was setup and made some things that weren't prefabs into prefabs.