Fully implemented pausing.
I didn't want to implement it this way but it's the best I could do for now. I had to do some goofy stuff to handle the coroutines and I'm honestly too lazy to write it all here. Essentially if a coroutine returns anything other than a WaitForFixedUpdate() it'll still execute normally. Therefore you have to implement a "Yield Return null" within a loop that's sentinel value is the isPaused var.