Update CharacterManager.cs
I started implementing the game mode and using that to detect if the player has won and also tracking which players are alive/not alive.

I also fixed a couple of issues with character selections where too many would be spawned or random errors were thrown.

I need to rework the CharacterManager class so it doesn't deal with any logic and only handles telling the GameMode when to remove a stock from a player.