Update CharacterManager.cs
This is still unfinished but I've almost got the UI for the Player/Character fully implemented such that when the player joins it shows them at the bottom of the selection screen and shows which character they've selected.

I just need to modify this so that when you join we only check from CharacterManager and instantiate the cursors manually that way it doesn't break every time you return to the selection screen.