Update CharacterManager.cs
I got the UI working on the character selection screen and it now remembers how many players already joined.

When players join it adds a new box for them.

I have not yet coded players leaving but I also need to change it so that any button will let a player join.

Because currently only actions that are marked as "Buttons" cause the player to join.

Which is weird because the space bar doesn't let you join even though it should be counted as an action. I assume that has something to do with the processor attached to that action.

I need to go eat and do homework though so we'll see if I continue working on this later.