Update CharacterManager.cs
I fixed the joining stuff so that when a player joins it only lets them join once for that device and still allows others to join.

I also fixed the issue where you couldn't place two coins on top of each other making it harder to choose a character.

Now you can.