Update CharacterManager.cs
These changes are from a week or so ago and I never uploaded them.

I mostly did a lot of stuff with the UI to make it work much better.

The character selection screen is a lot better now.

I still need to fix how the cursor is sized.

I added all the stuff for players joining and P1 - P8 always having the exact same color for each respectively (P1 is always a purplish color)