Update CharacterManager.cs & Cursor.cs
Changes:
Disconnecting Players are removed from the selection screen on Disconnect.

TODO:

In the future LD needs to finish coding something in the characterManager in the Disconnect section that pauses the match if the player disconnects and reassigns their controller to their player on reconnect.

LD needs to focus on finishing up some unfinished stuff in the player.